+++ src/cixl/types/sym.h
@@ -10,10 +10,10 @@ struct cx_type;
 
 struct cx_sym {
   char *id;
-  uint64_t tag;
+  size_t tag;
 };
 
-struct cx_sym *cx_sym_init(struct cx_sym *sym, const char *id, uint64_t tag);
+struct cx_sym *cx_sym_init(struct cx_sym *sym, const char *id, size_t tag);
 struct cx_sym *cx_sym_deinit(struct cx_sym *sym);
 
 enum cx_cmp cx_cmp_sym(const void *x, const void *y);
